+++ calculatormobile/pages/Login/Login.js
@@ -1,16 +1,16 @@
 import React from 'react';
-import { View, Text, SafeAreaView, StyleSheet, TouchableOpacity } from 'react-native';
+import { View, StyleSheet } from 'react-native';
 
 import LoginTypes from './components/LoginTypes';
 
 export default function Login() {
 	return (
-		<SafeAreaView style={styles.container}>
-			<View style={styles.titleContainer}>
-				<Text style={styles.title}>Login</Text>
+		<View style={styles.container}>
+			<View style={styles.user}></View>
+			<View style={styles.login__logout}>
+				<LoginTypes />
 			</View>
-			<LoginTypes />
-		</SafeAreaView>
+		</View>
 	);
 }
 
@@ -18,26 +18,12 @@ const styles = StyleSheet.create({
 	container: {
 		flex: 1,
 	},
-	titleContainer: {
-		flex: 1,
-		justifyContent: 'center',
-		alignItems: 'center',
-		backgroundColor: 'red',
+	user: {
+		flex: 8,
+		backgroundColor: 'lightgreen',
 	},
-	title: {
-		fontSize: 26,
-		fontWeight: 'bold',
-	},
-	closeBtnContainer: {
-		flex: 1,
-		justifyContent: 'center',
-		alignItems: 'center',
-	},
-	closeBtn: {
-		height: '30%',
-		width: '30%',
-		backgroundColor: 'gold',
-		justifyContent: 'center',
-		alignItems: 'center',
+	login__logout: {
+		backgroundColor: 'pink',
+		flex: 2,
 	},
 });
