+++ calculatormobile/pages/temp/EmailRegister.js
